Refactored findSignatureByUserId method
